Allow to override the http.Client for RequestToken and AccessToken
Before, the `http.DefaultClient` was always used.  This is e.g. not
sufficient, if the server being authenticated against uses TLS
certificates signed by a custom root CA.  In this case, a custom
`http.Client` needs to be passed.

This change follows the design of https://github.com/golang/oauth2
and allows passing the HTTP client via `context.Context`.

Since the user of a function that accepts a `context.Context` would
likely expect it also to be used for HTTP calls, we also pass it to
`http.NewRequest`.