Read body of error response
Previously a failing request to retrieve a request token or an access token with an unexpected status code was hard to debug, because additional information returned by the server in the body of the response was inaccessible to the user of this library.

This patch changes that by including the body in the error message. Since the body is now read before the potential status code error is returned, but it is crucial information for debugging, we also include the status code in an error resulting from reading the body.